requirements.txt
numpy
scikit-image
matplotlib
torch
torchvision
scikit-learn
Pillow
skorch
opencv-python